use size_t type instead of unsigned int for ft_strlen function

+++ ft_strlen.c
@@ -6,16 +6,18 @@
 /*   By: tbayrakt <tbayrakt@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/17 17:00:02 by tbayrakt          #+#    #+#             */
-/*   Updated: 2024/02/17 17:36:08 by tbayrakt         ###   ########.fr       */
+/*   Updated: 2024/02/18 11:42:11 by tbayrakt         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
+
+#include <stddef.h>
 
 // #include <stdio.h>
 // #include <string.h>
 
-unsigned int	ft_strlen(char *str)
+size_t	ft_strlen(char *str)
 {
-	int	counter;
+	size_t	counter;
 
 	counter = 0;
 	while (*str != '\0')
@@ -27,9 +29,9 @@ unsigned int	ft_strlen(char *str)
 }
 // int	main(void)
 // {
-// 	char word[10] = "";
+// 	char word[10] = "something";
 
-// 	printf("Length of word is: %d\n", ft_strlen(word));
+// 	printf("Length of word is: %zu\n", ft_strlen(word));
 // 	printf("Length of word in original function is: %lu\n", strlen(word));
 // 	return (0);
 // }
